add routes dir to store apis modularly, renamed API to userSearchByEmail)
