+++ src/components/CreateEvent.js
@@ -41,7 +41,7 @@ const CreateEvent = () => {
     }
 
     setEmail(currentUser.email);
-    loadUserEmailSearch(currentUser.email);
+    loaduserSearchByEmail(currentUser.email);
   }, []);
 
   //const email = currentUser.email;
@@ -132,8 +132,8 @@ const CreateEvent = () => {
     );
   });
 
-  const loadUserEmailSearch = (email) => {
-    callApiGetUserEmailSearch(email)
+  const loaduserSearchByEmail = (email) => {
+    callApiGetuserSearchByEmail(email)
       .then(res => {
         var parsed = JSON.parse(res.express);
         //console.log(parsed[0].id);
@@ -141,8 +141,8 @@ const CreateEvent = () => {
       });
   }
 
-  const callApiGetUserEmailSearch = async (email) => {
-    const url = `${REACT_APP_API_ENDPOINT}/userEmailSearch`;
+  const callApiGetuserSearchByEmail = async (email) => {
+    const url = `${REACT_APP_API_ENDPOINT}/userSearchByEmail`;
     console.log(url);
 
     const response = await fetch(url, {
